+++ src/modules/botinfo/commands/about.js
@@ -1,0 +1,16 @@
+export default {
+    args: '',
+    options: [],
+    /**
+     * Send bot's info.
+     * @param {Message} msg The universal msg object
+     */
+    async action(msg) {
+        msg.send(
+`Nodebot v${config.version}
+powered by Node.js & cqhttp.
+${new Date().getFullYear()} trustgit | under MIT License
+Star on GitLab if you enjoy it! https://gitlab.com/trustgit/nodebot`
+        )
+    }
+}